+++ libraries/AP_RCProtocol/AP_RCProtocol_Backend.h
@@ -19,6 +19,7 @@
 
 #include "AP_RCProtocol.h"
 #include <AP_HAL/utility/sparse-endian.h>
+#include <AP_VideoTX/AP_VideoTX_config.h>
 
 class AP_RCProtocol_Backend {
     friend class AP_RCProtcol;
@@ -91,8 +92,10 @@ public:
         return true;
     }
 
+#if AP_VIDEOTX_ENABLED
     // called by static methods to confiig video transmitters:
     static void configure_vtx(uint8_t band, uint8_t channel, uint8_t power, uint8_t pitmode);
+#endif
 
 protected:
